Introduce pedantic compiler flags by default

+++ src/asm.h
@@ -40,7 +40,7 @@ int print_registers(STREAM *out, const int32_t regs[16]);
 extern const struct format tenyr_asm_formats[];
 extern const size_t tenyr_asm_formats_count;
 
-int make_format_list(int (*pred)(const struct format *), size_t flen,
+size_t make_format_list(int (*pred)(const struct format *), size_t flen,
         const struct format *fmts, size_t len, char *buf,
         const char *sep);
