+++ src/asmif.c
@@ -240,7 +240,9 @@ static int fixup_deferred_exprs(struct parse_data *pd)
                 rc |= 1;
             }
 
-            uint32_t mask = ~((1LL << r->width) - 1);
+            // The mask may need to be all ones, so we need to compute the mask
+            // in a space larger than 32 bits, before truncating it.
+            uint32_t mask = (uint32_t)~((1ULL << r->width) - 1);
             *r->dest &= mask;
             *r->dest |= result & ~mask;
             ce_free(ce);
